requirements.txt
numpy
opencv-python
pillow
scikit-learn
tensorflow
imgaug
scipy
matplotlib
scikit-image
xgboost
seaborn
